[usrloc module]: added a new 'attr' column in the location table, used to store additional registration-related information
[registrar module]: new avp used to load/store the 'attr' column
information in the location table. Its content is read and stored in db and/or
memory when doing a 'save' operation, and it is populated after doing a 'lookup'.

When doing call forking, the avp will hold a value for each branch. The 'attr'
value of a branch is accessed using $T_branch_idx.

git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@9905 689a6050-402a-0410-94f2-e92a70836424

Conflicts:
	modules/usrloc/README